修复播放完成通知不生效

+++ app/src/main/cpp/media/MediaPlayer.h
@@ -26,7 +26,7 @@ private:
 public:
     Status *mStatus = NULL;
     double mClock = 0;
-    int mDuration = 0;
+    double mDuration = 0;
     int mRotation = 0;
     int mWidth = 0;
     int mHeight = 0;
